swords/requirements.txt
# Ticket Automation System - Python Dependencies
# Install: pip install -r requirements.txt

# GUI
# tkinter is built-in with Python

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Automation
pyautogui>=0.9.54

# HTTP Requests (for VLM API)
requests>=2.31.0

# Optional: NumPy (required by opencv)
numpy>=1.24.0
